update leaf handler and root validation
