+++ crates/event-watcher-traits/src/evm/event_watcher.rs
@@ -13,9 +13,6 @@
 // limitations under the License.
 
 use tokio::sync::Mutex;
-use webb::evm::contract::protocol_solidity::{
-    VAnchorContract, VAnchorContractEvents,
-};
 use webb_relayer_utils::retry;
 
 use super::*;
@@ -106,7 +103,10 @@ pub trait EventWatcher {
                 let src_typed_chain_id = TypedChainId::Evm(chain_id);
                 let history_store_key =
                     ResourceId::new(src_target_system, src_typed_chain_id);
-
+                let block = store.get_last_block_number(
+                    history_store_key,
+                    contract.deployed_at().as_u64(),
+                )?;
                 let current_block_number = client
                     .get_block_number()
                     .map_err(Into::into)
@@ -117,92 +117,6 @@ pub trait EventWatcher {
                     "Latest block number: #{}",
                     current_block_number
                 );
-                let block = store.get_last_block_number(
-                    history_store_key,
-                    contract.deployed_at().as_u64(),
-                )?;
-                tracing::trace!("Last saved block number: #{}", block);
-
-                let latest_merkel_root =
-                    store.get_latest_merkle_root(history_store_key)?;
-                let anchor_contract =
-                    VAnchorContract::new(contract.address(), client.clone());
-                let on_chain_root: [u8; 32] = anchor_contract
-                    .get_last_root()
-                    .call()
-                    .map_err(Into::into)
-                    .map_err(backoff::Error::transient)
-                    .await?
-                    .into();
-
-                if latest_merkel_root == on_chain_root.to_vec() {
-                    tracing::info!(
-                        "Computed merkle roots matches with on chain merkle root"
-                    );
-                } else {
-                    tracing::info!("Fast sync blocks");
-                    let interval = 1000;
-                    let contract_deployed_at = contract.deployed_at().as_u64();
-                    let mut relayer_leaves: Vec<[u8; 32]> = Vec::new();
-                    for block_no in (contract_deployed_at
-                        ..=current_block_number)
-                        .step_by(interval)
-                    {
-                        let from_block = block_no as u64;
-                        let to_block = block_no + interval as u64 - 1;
-                        tracing::trace!(
-                            "Fast sync blocks from #{} to #{}",
-                            from_block,
-                            to_block
-                        );
-                        let events_filter = contract
-                            .event_with_filter::<VAnchorContractEvents>(
-                                Default::default(),
-                            )
-                            .from_block(from_block)
-                            .to_block(to_block);
-                        let found_events = events_filter
-                            .query_with_meta()
-                            .map_err(Into::into)
-                            .map_err(backoff::Error::transient)
-                            .await?;
-                        for (event, _log) in found_events {
-                            match event {
-                                VAnchorContractEvents::NewCommitmentFilter(
-                                    deposit,
-                                ) => {
-                                    let commitment: [u8; 32] =
-                                        deposit.commitment.into();
-                                    relayer_leaves.push(commitment);
-                                }
-                                _ => {
-                                    tracing::trace!(
-                                        "Unhandled event {:?}",
-                                        event
-                                    );
-                                }
-                            }
-                        }
-                    }
-                    tracing::info!("Fast sync completed");
-
-                    // compute root using leaves fetched from blocks
-                    let computed_root = compute_merkle_root(relayer_leaves);
-                    if computed_root == on_chain_root {
-                        // save computed merkle root
-                        store.set_latest_merkle_root(
-                            history_store_key,
-                            computed_root,
-                        );
-                        // TODO!
-                        // save last processed block number
-                        // clear cache
-                        // write cache
-                        // relay roots
-                        continue;
-                    }
-                }
-
                 // latest finalized block after n block_confirmations
                 let latest_finalized_block =
                     current_block_number.saturating_sub(block_confirmations);
@@ -323,7 +237,6 @@ pub trait EventWatcher {
         Ok(())
     }
 }
-
 /// A trait that defines a handler for a specific set of event types.
 ///
 /// The handlers are implemented separately from the watchers, so that we can have
@@ -354,6 +267,13 @@ pub trait EventHandler {
         (event, log): (Self::Events, contract::LogMeta),
         metrics: Arc<Mutex<metric::Metrics>>,
     ) -> webb_relayer_utils::Result<()>;
+
+    /// Whether any of the events could be handled by the handler
+    async fn can_handle_events(
+        &self,
+        event: Self::Events,
+        wrapper: &Self::Contract,
+    ) -> webb_relayer_utils::Result<bool>;
 }
 
 /// An Auxiliary trait to handle events with retry logic.
@@ -379,6 +299,10 @@ pub trait EventHandlerWithRetry: EventHandler {
         backoff: impl backoff::backoff::Backoff + Send + Sync + 'static,
         metrics: Arc<Mutex<metric::Metrics>>,
     ) -> webb_relayer_utils::Result<()> {
+        if !self.can_handle_events(event.clone(), contract).await? {
+            return Ok(());
+        };
+
         let wrapped_task = || {
             self.handle_event(
                 store.clone(),
